requirements.txt
gradio>=5.0.0
rapidocr-onnxruntime
Pillow
numpy
opencv-python-headless
paddle2onnx
